updated tab placeholder On branch master Your branch is up-to-date with 'origin/master'. Changes to be committed:   (use "git reset HEAD <file>..." to unstage)
	modified:   stylesheets/tabs.less Please enter the commit message for your changes. Lines starting
On branch master
Your branch is up-to-date with 'origin/master'.
Changes to be committed:
  (use "git reset HEAD <file>..." to unstage)

	modified:   stylesheets/tabs.less with 'On branch master
Your branch is up-to-date with 'origin/master'.
Changes to be committed:
  (use "git reset HEAD <file>..." to unstage)

	modified:   stylesheets/tabs.less' will be ignored, and an empty message aborts the commit.
On branch master
Your branch is up-to-date with 'origin/master'.
Changes to be committed:
  (use "git reset HEAD <file>..." to unstage)

	modified:   stylesheets/tabs.less
On branch master
Your branch is up-to-date with 'origin/master'.
Changes to be committed:
  (use "git reset HEAD <file>..." to unstage)

	modified:   stylesheets/tabs.less On branch master
On branch master
Your branch is up-to-date with 'origin/master'.
Changes to be committed:
  (use "git reset HEAD <file>..." to unstage)

	modified:   stylesheets/tabs.less Your branch is up-to-date with 'origin/master'.
On branch master
Your branch is up-to-date with 'origin/master'.
Changes to be committed:
  (use "git reset HEAD <file>..." to unstage)

	modified:   stylesheets/tabs.less Changes to be committed:
On branch master
Your branch is up-to-date with 'origin/master'.
Changes to be committed:
  (use "git reset HEAD <file>..." to unstage)

	modified:   stylesheets/tabs.less 	modified:   stylesheets/tabs.less